Make the example smaller, developer focused
