+++ README.md
@@ -29,14 +29,13 @@ Agent](https://docs.splunk.com/Observability/gdi/get-data-in/application/go/trou
 Table of Contents:
 
 - [Getting started](#getting-started)
-  - [Basic configuration](#basic-configuration)
+	- [Basic configuration](#basic-configuration)
+	- [Examples](#examples)
+- [Advanced configuration](#advanced-configuration)
+- [Correlate traces and logs](#correlate-traces-and-logs)
 - [Library instrumentation](#library-instrumentation)
 - [Manual instrumentation](#manual-instrumentation)
-- [Advanced configuration](#advanced-configuration)
-  - [Splunk distribution configuration](#splunk-distribution-configuration)
-  - [Trace configuration](#trace-configuration)
-  - [Trace exporter configuration](#trace-exporter-configuration)
-  - [Trace propagation configuration](#trace-propagation-configuration)
+- [Troubleshooting](#troubleshooting)
 - [License](#license)
 
 ## Getting started
@@ -106,6 +105,14 @@ os.Setenv("OTEL_RESOURCE_ATTRIBUTES", "service.name=my-app,service.version=1.2.3
 ```
 
 For advanced configuration options, refer to the [`distro` package documentation](./distro/README.md#Configuration).
+
+### Examples
+
+You can find our official "user-facing" examples
+[here](https://github.com/signalfx/tracing-examples/tree/main/opentelemetry-tracing/opentelemetry-go).
+
+The [example](./example) contains a smaller, developer focused, example.
+It is meant to be used for experimenting and exploratory end-to-end testing.
 
 ## Advanced configuration
