--- example/clean.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
-
-docker-compose -f "${DIR}/docker-compose.yaml" down
