--- example/client/go.mod
@@ -1,10 +1,0 @@
-module client
-
-go 1.16
-
-require (
-	github.com/signalfx/splunk-otel-go/distro v0.8.0
-	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.31.0
-)
-
-replace github.com/signalfx/splunk-otel-go/distro => ../../distro
