--- example/client/main.go
@@ -1,77 +1,0 @@
-// Copyright Splunk Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"os"
-	"os/signal"
-	"time"
-
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-
-	"github.com/signalfx/splunk-otel-go/distro"
-)
-
-func main() {
-	// handle CTRL+C gracefully
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt)
-	defer stop()
-
-	// initialize Splunk OTel distro
-	sdk, err := distro.Run()
-	if err != nil {
-		panic(err)
-	}
-	defer func() {
-		if err := sdk.Shutdown(context.Background()); err != nil {
-			panic(err)
-		}
-	}()
-
-	// instrument http.Client
-	client := &http.Client{Transport: otelhttp.NewTransport(http.DefaultTransport)}
-
-	for {
-		select {
-		case <-ctx.Done():
-			stop() // stop receiving signal notifications; next interrupt signal should kill the application
-			return
-		case <-time.After(time.Second):
-			call(ctx, client)
-		}
-	}
-}
-
-func call(ctx context.Context, client *http.Client) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, "http://localhost:8080/hello", http.NoBody)
-	if err != nil {
-		panic(err)
-	}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Println(err)
-		return
-	}
-	defer resp.Body.Close()
-	if _, err := io.Copy(os.Stdout, resp.Body); err != nil {
-		log.Println(err)
-	}
-	fmt.Printf(" HTTP Headers: %v\n", resp.Header)
-}
