+++ example/docker-compose.yaml
@@ -1,31 +1,7 @@
 version: "3.9"
 services:
-  server:
-    image: golang:1.18
-    volumes: [ ../:/repo ]
-    working_dir: /repo/example/server
-    entrypoint: go run .
-    environment:
-      - OTEL_SERVICE_NAME=server
-      - OTEL_RESOURCE_ATTRIBUTES
-    network_mode: host
-    depends_on:
-      - otel-collector
-
-  client:
-    image: golang:1.18
-    volumes: [ ../:/repo ]
-    working_dir: /repo/example/client
-    entrypoint: go run .
-    environment:
-      - OTEL_SERVICE_NAME=client
-      - OTEL_RESOURCE_ATTRIBUTES
-    network_mode: host
-    depends_on:
-      - otel-collector
-
   otel-collector:
-    image: otel/opentelemetry-collector-contrib:0.47.0
+    image: otel/opentelemetry-collector-contrib:0.49.0
     volumes:
       - ./otel-config.yaml:/etc/otel/config.yaml
     command: --config /etc/otel/config.yaml
@@ -47,5 +23,4 @@ services:
     image: jaegertracing/all-in-one:latest
     ports:
       - "16686:16686"
-      - "14268:14268"
       - "14250:14250"
